Swap info + cancel order, fixbug trello 20/02 (#62)
* Remove core data

* https://trello.com/c/0AYW4WsL/112-ui-swap-settings-have-an-issue-when-tap-on-the-settings-icon

* https://trello.com/c/pdUGeNNh/113-ui-orders-order-detail-missing-expires-at-warning-and-expires-at-field-in-the-execution-information-column

* Update swap info

* Add api ios trade

* Update logic

* Update price impact

* Update logic

* Update logic swap

* Add cancel order

* Add cancel order

* Cancel order detail

* Update swap token

* Update trade

* Update code

* Update tool tip

* https://trello.com/c/WWqQw00H/117-ui-token-detail-the-interval-time-bar

* Update lint

* Update swap

* Update swap

* Update logic swap

* Update logic

* Update token

* Update alert

* Update UI

* Update text field

* Update logic

* Update code

* Update swap

* Update swap

* Refactor

+++ MinWallet/Features/Home/SearchTokenView.swift
@@ -21,6 +21,9 @@ struct SearchTokenView: View {
                         .placeholder("Search", when: viewModel.keyword.isEmpty)
                         .focused($isFocus)
                         .lineLimit(1)
+                        .keyboardType(.asciiCapable)
+                        .submitLabel(.done)
+                        .autocorrectionDisabled()
                     if !viewModel.keyword.isBlank {
                         Image(.icCloseFill)
                             .fixSize(16)
@@ -77,6 +80,7 @@ struct SearchTokenView: View {
                 Spacer(minLength: 0)
             }
         }
+        /*
         .toolbar {
             ToolbarItemGroup(placement: .keyboard) {
                 Spacer()
@@ -86,6 +90,7 @@ struct SearchTokenView: View {
                 .foregroundStyle(.colorLabelToolbarDone)
             }
         }
+         */
         .background(.colorBaseBackground)
         .onAppear {
             viewModel.getTokens()
@@ -138,38 +143,39 @@ struct SearchTokenView: View {
                 .frame(height: 32)
                 .padding(.horizontal, .xl)
             ForEach(0..<viewModel.tokensFav.count, id: \.self) { index in
-                let item = viewModel.tokensFav[index]
-                let offsetBinding = Binding<CGFloat>(
-                    get: {
-                        viewModel.offsets[gk_safeIndex: index] ?? 0
-                    },
-                    set: { value in
-                        guard index >= 0, index < viewModel.offsets.count else { return }
-                        viewModel.offsets[index] = value
-                    }
-                )
-                let deleteBinding = Binding<Bool>(
-                    get: {
-                        viewModel.isDeleted[gk_safeIndex: index] ?? false
-                    },
-                    set: { value in
-                        guard index >= 0, index < viewModel.isDeleted.count else { return }
-                        viewModel.isDeleted[index] = value
-                    }
-                )
-
-                TokenListItemView(token: item, showBottomLine: index != viewModel.tokensFav.count - 1)
-                    .contentShape(.rect)
-                    .swipeToDelete(
-                        offset: offsetBinding, isDeleted: deleteBinding, height: 68,
-                        onDelete: {
-                            viewModel.deleteTokenFav(at: index)
+                if let item = viewModel.tokensFav[gk_safeIndex: index] {
+                    let offsetBinding = Binding<CGFloat>(
+                        get: {
+                            viewModel.offsets[gk_safeIndex: index] ?? 0
+                        },
+                        set: { value in
+                            guard index >= 0, index < viewModel.offsets.count else { return }
+                            viewModel.offsets[index] = value
                         }
                     )
-                    .zIndex(Double(index) * -1)
-                    .onTapGesture {
-                        navigator.push(.tokenDetail(token: item))
-                    }
+                    let deleteBinding = Binding<Bool>(
+                        get: {
+                            viewModel.isDeleted[gk_safeIndex: index] ?? false
+                        },
+                        set: { value in
+                            guard index >= 0, index < viewModel.isDeleted.count else { return }
+                            viewModel.isDeleted[index] = value
+                        }
+                    )
+
+                    TokenListItemView(token: item, showBottomLine: index != viewModel.tokensFav.count - 1)
+                        .contentShape(.rect)
+                        .swipeToDelete(
+                            offset: offsetBinding, isDeleted: deleteBinding, height: 68,
+                            onDelete: {
+                                viewModel.deleteTokenFav(at: index)
+                            }
+                        )
+                        .zIndex(Double(index) * -1)
+                        .onTapGesture {
+                            navigator.push(.tokenDetail(token: item))
+                        }
+                }
             }
         }
     }
@@ -196,16 +202,17 @@ struct SearchTokenView: View {
                 .padding(.top, spacing)
         }
         ForEach(0..<viewModel.tokens.count, id: \.self) { index in
-            let item = viewModel.tokens[index]
-            TokenListItemView(token: item)
-                .contentShape(.rect)
-                .onTapGesture {
-                    viewModel.addRecentSearch(keyword: item.adaName)
-                    navigator.push(.tokenDetail(token: item))
-                }
-                .onAppear() {
-                    viewModel.loadMoreData(item: item)
-                }
+            if let item = viewModel.tokens[gk_safeIndex: index] {
+                TokenListItemView(token: item)
+                    .contentShape(.rect)
+                    .onTapGesture {
+                        viewModel.addRecentSearch(keyword: item.adaName)
+                        navigator.push(.tokenDetail(token: item))
+                    }
+                    .onAppear() {
+                        viewModel.loadMoreData(item: item)
+                    }
+            }
         }
     }
 }
